refactor(tools): split records out of test_runs
- Test run and test record tools shared one 898-line module despite
  disjoint resources, guards, models, and parse helpers; guard layer
  already keeps the two domains in separate modules.
- Move the four record tools with their payload builders into
  tools/test_records.py and split the test mirror accordingly; tool
  names, behavior, and descriptions unchanged.
