+++ src/mcp_server_polarion/tools/__init__.py
@@ -12,6 +12,7 @@ import mcp_server_polarion.tools.links
 import mcp_server_polarion.tools.moves
 import mcp_server_polarion.tools.projects
 import mcp_server_polarion.tools.recipes
+import mcp_server_polarion.tools.test_records
 import mcp_server_polarion.tools.test_runs
 import mcp_server_polarion.tools.work_items  # noqa: F401
